Rename api.py to views.py, fix Status regression
