+++ example/tests/basket/test_basket_views.py
@@ -3,7 +3,7 @@ from django.urls import reverse
 from rest_framework.settings import api_settings
 from rest_framework.test import APIClient
 
-from salesman.basket.api import BasketViewSet
+from salesman.basket.views import BasketViewSet
 from salesman.core.utils import get_salesman_model
 from shop.models import Product
 
@@ -11,7 +11,7 @@ BasketItem = get_salesman_model('BasketItem')
 
 
 @pytest.mark.django_db
-def test_basket_api():
+def test_basket_views():
     url = reverse('salesman-basket-list')
     client = APIClient()
