+++ example/tests/checkout/test_checkout_views.py
@@ -2,7 +2,7 @@ import pytest
 from django.urls import reverse
 from rest_framework.test import APIClient
 
-from salesman.checkout.api import CheckoutViewSet
+from salesman.checkout.views import CheckoutViewSet
 from salesman.core.utils import get_salesman_model
 from shop.models import Product
 
@@ -10,7 +10,7 @@ Order = get_salesman_model('Order')
 
 
 @pytest.mark.django_db
-def test_checkout_api(settings):
+def test_checkout_views(settings):
     url = reverse('salesman-checkout-list')
     client = APIClient()
