+++ salesman/urls.py
@@ -1,9 +1,9 @@
 from rest_framework.routers import DefaultRouter
 
-from salesman.basket.api import BasketViewSet
-from salesman.checkout.api import CheckoutViewSet
+from salesman.basket.views import BasketViewSet
 from salesman.checkout.payment import payment_methods_pool
-from salesman.orders.api import OrderViewSet
+from salesman.checkout.views import CheckoutViewSet
+from salesman.orders.views import OrderViewSet
 
 router = DefaultRouter()
 router.register('basket', BasketViewSet, basename='salesman-basket')
